feat: improvements — clickable records, shared module, build fix
- Delete unused expandable-grants-table.tsx
- Extract shared grant components to grants/grant-shared.tsx (DRY)
- Update org grant detail page to use shared module
- Make funding round names clickable (→ /funding-rounds/[id])
- Make investment round names clickable (→ /investments/[id])
- Update funding-programs grants to use org-scoped URLs
- Fix factbase build: add @longterm-wiki/factbase to transpilePackages
