+++ apps/web/src/app/organizations/[slug]/grants/[grantId]/page.tsx
@@ -3,10 +3,8 @@ import type { Metadata } from "next";
 import Link from "next/link";
 import {
   getAllKBRecords,
-  getKBEntity,
   getKBEntitySlug,
 } from "@/data/factbase";
-import type { KBRecordEntry } from "@/data/factbase";
 import { getTypedEntityById, getRecordVerdict } from "@/data/tablebase";
 import { formatCompactCurrency } from "@/lib/format-compact";
 import { Breadcrumbs } from "@/components/directory";
@@ -20,80 +18,12 @@ import {
 } from "@/components/wiki/factbase/format";
 import { resolveOrgBySlug, getOrgSlugs } from "@/app/organizations/org-utils";
 import { STATUS_COLORS } from "@/app/grants/grants-constants";
-
-// ── Types ──────────────────────────────────────────────────────────────
-
-interface ParsedGrant {
-  key: string;
-  ownerEntityId: string;
-  name: string;
-  funderName: string;
-  funderSlug: string | null;
-  funderHref: string | null;
-  recipientId: string | null;
-  recipientName: string;
-  recipientHref: string | null;
-  amount: number | null;
-  currency: string | null;
-  date: string | null;
-  period: string | null;
-  status: string | null;
-  source: string | null;
-  program: string | null;
-  programId: string | null;
-  notes: string | null;
-}
-
-// ── Resolution helpers ─────────────────────────────────────────────────
-
-function resolveEntityLink(entityId: string): {
-  name: string;
-  slug: string | null;
-  href: string | null;
-} {
-  const entity = getKBEntity(entityId);
-  if (entity) {
-    const slug = getKBEntitySlug(entityId);
-    if (slug) {
-      if (entity.type === "organization")
-        return { name: entity.name, slug, href: `/organizations/${slug}` };
-      if (entity.type === "person")
-        return { name: entity.name, slug, href: `/people/${slug}` };
-    }
-    return { name: entity.name, slug: null, href: `/factbase/entity/${entityId}` };
-  }
-  return { name: titleCase(entityId.replace(/-/g, " ")), slug: null, href: null };
-}
-
-function parseGrant(record: KBRecordEntry): ParsedGrant {
-  const f = record.fields;
-  const funder = resolveEntityLink(record.ownerEntityId);
-  const recipientId = typeof f.recipient === "string" ? f.recipient : null;
-  const recipient = recipientId
-    ? resolveEntityLink(recipientId)
-    : { name: "", slug: null, href: null };
-
-  return {
-    key: record.key,
-    ownerEntityId: record.ownerEntityId,
-    name: (f.name as string) ?? record.key,
-    funderName: funder.name,
-    funderSlug: funder.slug,
-    funderHref: funder.href,
-    recipientId,
-    recipientName: recipient.name,
-    recipientHref: recipient.href,
-    amount: typeof f.amount === "number" ? f.amount : null,
-    currency: typeof f.currency === "string" ? f.currency : null,
-    date: typeof f.date === "string" ? f.date : null,
-    period: typeof f.period === "string" ? f.period : null,
-    status: typeof f.status === "string" ? f.status : null,
-    source: typeof f.source === "string" ? f.source : null,
-    program: typeof f.program === "string" ? f.program : null,
-    programId: typeof f.programId === "string" ? f.programId : null,
-    notes: typeof f.notes === "string" ? f.notes : null,
-  };
-}
+import {
+  parseGrantDetail,
+  DetailSection,
+  EntityLinkDisplay,
+  RelatedGrantsSection,
+} from "@/app/grants/grant-shared";
 
 // ── Static params ─────────────────────────────────────────────────────
 
@@ -128,7 +58,7 @@ export async function generateMetadata({ params }: PageProps): Promise<Metadata>
   const funderSlug = getKBEntitySlug(record.ownerEntityId);
   if (funderSlug !== slug) return { title: "Grant Not Found" };
 
-  const grant = parseGrant(record);
+  const grant = parseGrantDetail(record);
   const org = resolveOrgBySlug(slug);
   const orgName = org?.name ?? slug;
   const parts = [grant.name];
@@ -149,13 +79,12 @@ export default async function OrgGrantDetailPage({ params }: PageProps) {
 
   if (!record) notFound();
 
-  const grant = parseGrant(record);
+  const grant = parseGrantDetail(record);
   const grantVerdict = getRecordVerdict("grant", String(grant.key));
 
   // Verify this grant belongs to the org (by funder slug)
   const funderSlug = getKBEntitySlug(record.ownerEntityId);
   if (funderSlug !== slug) {
-    // Grant exists but belongs to a different org — redirect
     if (funderSlug) {
       redirect(`/organizations/${funderSlug}/grants/${grantId}`);
     }
@@ -165,7 +94,7 @@ export default async function OrgGrantDetailPage({ params }: PageProps) {
   // Find related grants: same funder
   const relatedByFunder = allGrants
     .filter((r) => r.ownerEntityId === grant.ownerEntityId && r.key !== grant.key)
-    .map(parseGrant);
+    .map(parseGrantDetail);
 
   // Find related grants: same recipient (from other funders)
   const relatedByRecipient = grant.recipientId
@@ -177,7 +106,7 @@ export default async function OrgGrantDetailPage({ params }: PageProps) {
             r.fields.recipient === grant.recipientId &&
             r.ownerEntityId !== grant.ownerEntityId,
         )
-        .map(parseGrant)
+        .map(parseGrantDetail)
     : [];
 
   // Funder wiki page link
@@ -345,130 +274,5 @@ export default async function OrgGrantDetailPage({ params }: PageProps) {
         </Link>
       </div>
     </div>
-  );
-}
-
-// ── Subcomponents ──────────────────────────────────────────────────────
-
-function DetailSection({
-  title,
-  children,
-}: {
-  title: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <div>
-      <div className="text-[11px] font-medium uppercase tracking-wider text-muted-foreground/70 mb-1">
-        {title}
-      </div>
-      <div className="flex items-center gap-1 flex-wrap">{children}</div>
-    </div>
-  );
-}
-
-function EntityLinkDisplay({
-  name,
-  href,
-}: {
-  name: string;
-  href: string | null;
-}) {
-  if (href) {
-    return (
-      <Link
-        href={href}
-        className="text-sm font-medium text-primary hover:underline"
-      >
-        {name}
-      </Link>
-    );
-  }
-  return <span className="text-sm font-medium text-foreground">{name}</span>;
-}
-
-function RelatedGrantsSection({
-  title,
-  grants,
-  totalCount,
-  orgSlug,
-}: {
-  title: string;
-  grants: ParsedGrant[];
-  totalCount: number;
-  orgSlug?: string;
-}) {
-  return (
-    <section className="mb-8">
-      <div className="flex items-center gap-3 mb-4">
-        <h2 className="text-base font-bold tracking-tight">{title}</h2>
-        <span className="text-[11px] font-medium tabular-nums px-2 py-0.5 rounded-full bg-muted text-muted-foreground">
-          {totalCount}
-        </span>
-        <div className="flex-1 h-px bg-gradient-to-r from-border/60 to-transparent" />
-      </div>
-      <div className="border border-border/60 rounded-xl overflow-x-auto">
-        <table className="w-full text-sm">
-          <thead>
-            <tr className="text-xs text-muted-foreground border-b border-border bg-muted/30">
-              <th className="text-left py-2 px-3 font-medium">Grant</th>
-              <th className="text-left py-2 px-3 font-medium">Recipient</th>
-              <th className="text-right py-2 px-3 font-medium">Amount</th>
-              <th className="text-center py-2 px-3 font-medium">Date</th>
-            </tr>
-          </thead>
-          <tbody className="divide-y divide-border/50">
-            {grants.map((g) => {
-              const grantSlug = g.funderSlug ?? orgSlug;
-              const href = grantSlug
-                ? `/organizations/${grantSlug}/grants/${g.key}`
-                : null;
-              return (
-                <tr key={g.key} className="hover:bg-muted/20 transition-colors">
-                  <td className="py-2 px-3">
-                    {href ? (
-                      <Link
-                        href={href}
-                        className="font-medium text-foreground text-xs hover:text-primary transition-colors"
-                      >
-                        {g.name}
-                      </Link>
-                    ) : (
-                      <span className="font-medium text-foreground text-xs">
-                        {g.name}
-                      </span>
-                    )}
-                  </td>
-                  <td className="py-2 px-3 text-xs">
-                    {g.recipientHref ? (
-                      <Link href={g.recipientHref} className="text-primary hover:underline">
-                        {g.recipientName}
-                      </Link>
-                    ) : (
-                      <span className="text-muted-foreground">{g.recipientName}</span>
-                    )}
-                  </td>
-                  <td className="py-2 px-3 text-right tabular-nums whitespace-nowrap text-xs">
-                    {g.amount != null && (
-                      <span className="font-semibold">
-                        {formatCompactCurrency(g.amount)}
-                      </span>
-                    )}
-                  </td>
-                  <td className="py-2 px-3 text-center text-muted-foreground text-xs">
-                    {g.date ? formatKBDate(g.date) : g.period ?? ""}
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
-      {totalCount > 10 && (
-        <div className="mt-2 text-xs text-muted-foreground text-center">
-          Showing 10 of {totalCount} grants
-        </div>
-      )}
-    </section>
   );
 }
